all works fine

+++ QtMy/QAxisSlider.h
@@ -10,7 +10,7 @@ public:
 	void bindSliderAndPointAxis(float* v,bool* nu)
 	{
 		axis_value_ = v;
-		*v = value();
+		*v = static_cast<float>(value());
 		bezier_need_update_ = nu;
 	}
 
@@ -18,7 +18,6 @@ protected:
 	void sliderChange(SliderChange change) override
 	{
 		QSlider::sliderChange(change);
-
 		if (change == SliderValueChange && axis_value_!=nullptr)
 		{
 			*axis_value_ = static_cast<float>(value());
